requirements.txt
beautifulsoup4==4.9.1
mysql-connector-python==8.0.21
requests==2.24.0